pyproject.toml
[project]
name = "openhumanoid"
version = "0.1.0"
description = "Voice-controlled humanoid robot integrating OpenClaw, DiMOS, and GR00T VLA"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "sounddevice>=0.4",
    "numpy>=1.24",
    "requests>=2.31",
    "python-dotenv>=1.0",
]
